Fix rate limiting
