+++ cognee/infrastructure/databases/vector/embeddings/LiteLLMEmbeddingEngine.py
@@ -10,7 +10,7 @@ from cognee.infrastructure.llm.tokenizer.Gemini import GeminiTokenizer
 from cognee.infrastructure.llm.tokenizer.HuggingFace import HuggingFaceTokenizer
 from cognee.infrastructure.llm.tokenizer.Mistral import MistralTokenizer
 from cognee.infrastructure.llm.tokenizer.TikToken import TikTokenTokenizer
-from cognee.infrastructure.llm.rate_limiter import (
+from cognee.infrastructure.llm.embedding_rate_limiter import (
     embedding_rate_limit_async,
     embedding_sleep_and_retry_async,
 )
